# prevent unnecessary recursion during trace