--- a/lib/trace.js
+++ b/lib/trace.js
@@ -628,6 +628,11 @@ Trace.prototype.getAllLoadRecords = function(canonical, traceOpts, canonicalCond
   var self = this;
   return this.getLoadRecord(canonical, traceOpts, parentStack)
   .then(function(load) {
+
+    // check this again – might have been loaded asynchronously since the last check
+    if (canonical in curLoads)
+      return curLoads;
+
     // conditionals, build: false and system modules are falsy loads in the trace trees
     // (that is, part of depcache, but not built)
     // we skip system modules though