Fixes #701 

The current exit condition is sub-optimal. As the tree of load dependencies is traversed asynchronously, it is possible that the same module's dependencies could be recursed many times. In our case, this was triggering a massive number of promises and crashing node itself.