@ghempton 2016-10-01
FWIW, this also gives us a *massive* performance boost locally:

Before:

```
./bin/build  146.57s user 6.64s system 101% cpu 2:31.69 total
```

After:

```
./bin/build  54.98s user 4.35s system 101% cpu 58.374 total
```