@guybedford 2016-10-02
getLoadRecord itself does have a synchronous queue through the `tracing` object at https://github.com/systemjs/builder/blob/master/lib/trace.js#L140.

But this will avoid duplicate tracing work on the tree certainly, which would be especially important with deep module trees.

Thanks so much for tracking this down!